Refactor fixed timestep handling and add bounds checks
- Introduced a new `fixed_step` function to normalize fixed timestep values, ensuring a minimum duration of 1ms and a maximum of 1s.
- Updated the `run_runtime` function to utilize the new `fixed_step` function for calculating the fixed update duration.
- Added unit tests for `fixed_step` to validate behavior for extreme rates and ensure bounded non-zero steps.
- Enhanced the `normalize_fixed_timestep_seconds` function to cap values at 1000 Hz and handle non-finite inputs gracefully.
- Implemented checks in the scene loader to limit the maximum number of custom parameters and instances to prevent potential overflows and memory issues.
- Added tests for parameter index limits and instance grid limits to ensure robustness against excessive values.
- Improved error handling in the lexer for invalid escape sequences in strings.
- Updated the parser to enforce a maximum nesting depth for typed blocks to prevent stack overflows.
- Conducted a thorough bug audit and addressed identified issues, including potential hangs and memory blowups in various components.

+++ perro_source/audio_stack/perro_pawdio/src/player.rs
@@ -58,7 +58,7 @@ impl BarkPlayer {
     // plays skip the decoder; longer clips stream-decode per play.
     const PCM_CACHE_MAX_SECONDS: usize = 12;
     const CACHE_EVICT_SWEEP_INTERVAL: Duration = Duration::from_millis(100);
-    const UNRESERVED_TTL_FACTOR: f32 = 2.0;
+    const UNRESERVED_TTL_FACTOR: u32 = 2;
     const UNRESERVED_TTL_FALLBACK: Duration = Duration::from_secs(1);
     const UNRESERVED_TTL_MIN: Duration = Duration::from_millis(250);
